Split config_vrouter_test from config_test
