+++ pkg/apis/contrail/v1alpha1/tests/config_test.go
@@ -716,56 +716,6 @@ func TestWebuiConfig(t *testing.T) {
 	}
 }
 
-type vrouterClusterInfoFake struct {
-	clusterName             string
-	cniBinariesDirectory    string
-	cniConfigFilesDirectory string
-}
-
-func (c vrouterClusterInfoFake) KubernetesClusterName() (string, error) {
-	return c.clusterName, nil
-}
-
-func (c vrouterClusterInfoFake) CNIBinariesDirectory() string {
-	return c.cniBinariesDirectory
-}
-func (c vrouterClusterInfoFake) CNIConfigFilesDirectory() string {
-	return c.cniConfigFilesDirectory
-}
-
-func TestVrouterConfig(t *testing.T) {
-	logf.SetLogger(logf.ZapLogger(true))
-
-	environment := SetupEnv()
-	cl := *environment.client
-
-	err := environment.vrouterResource.InstanceConfiguration(reconcile.Request{types.NamespacedName{Name: "vrouter1", Namespace: "default"}},
-		&environment.vrouterPodList, cl, vrouterClusterInfoFake{clusterName: "test-cluster"})
-	if err != nil {
-		t.Fatalf("get configmap: (%v)", err)
-	}
-	err = cl.Get(context.TODO(),
-		types.NamespacedName{Name: "vrouter1-vrouter-configmap", Namespace: "default"},
-		&environment.vrouterConfigMap)
-	if err != nil {
-		t.Fatalf("get configmap: (%v)", err)
-	}
-	err = cl.Get(context.TODO(),
-		types.NamespacedName{Name: "vrouter1-vrouter-configmap-1", Namespace: "default"},
-		&environment.vrouterConfigMap2)
-	if err != nil {
-		t.Fatalf("get configmap: (%v)", err)
-	}
-	if environment.vrouterConfigMap.Data["vrouter.1.1.8.1"] != vrouterConfig {
-		configDiff := diff.Diff(environment.vrouterConfigMap.Data["vrouter.1.1.8.1"], vrouterConfig)
-		t.Fatalf("get vrouter config: \n%v\n", configDiff)
-	}
-
-	if environment.vrouterConfigMap.Data["10-contrail.conf"] != vrouterCniConfig {
-		configDiff := diff.Diff(environment.vrouterConfigMap.Data["10-contrail.conf"], vrouterCniConfig)
-		t.Fatalf("get vrouter cni config: \n%v\n", configDiff)
-	}
-}
 
 func TestCassandraConfig(t *testing.T) {
 	logf.SetLogger(logf.ZapLogger(true))
@@ -1761,69 +1711,6 @@ sandesh_keyfile=/etc/certificates/server-key-1.1.6.1.pem
 sandesh_certfile=/etc/certificates/server-1.1.6.1.crt
 sandesh_ca_cert=/etc/ssl/certs/kubernetes/ca-bundle.crt`
 
-var vrouterConfig = `[CONTROL-NODE]
-servers=1.1.5.1:5269 1.1.5.2:5269 1.1.5.3:5269
-[DEFAULT]
-http_server_ip=0.0.0.0
-collectors=1.1.1.1:8086 1.1.1.2:8086 1.1.1.3:8086
-log_file=/var/log/contrail/contrail-vrouter-agent.log
-log_level=SYS_NOTICE
-log_local=1
-hostname=host1
-agent_name=host1
-xmpp_dns_auth_enable=True
-xmpp_auth_enable=True
-xmpp_server_cert=/etc/certificates/server-1.1.8.1.crt
-xmpp_server_key=/etc/certificates/server-key-1.1.8.1.pem
-xmpp_ca_cert=/etc/ssl/certs/kubernetes/ca-bundle.crt
-physical_interface_mac = de:ad:be:ef:ba:be
-tsn_servers = []
-[SANDESH]
-introspect_ssl_enable=True
-introspect_ssl_insecure=False
-sandesh_ssl_enable=True
-sandesh_keyfile=/etc/certificates/server-key-1.1.8.1.pem
-sandesh_certfile=/etc/certificates/server-1.1.8.1.crt
-sandesh_ca_cert=/etc/ssl/certs/kubernetes/ca-bundle.crt
-[NETWORKS]
-control_network_ip=1.1.8.1
-[DNS]
-servers=1.1.5.1:53 1.1.5.2:53 1.1.5.3:53
-[METADATA]
-metadata_proxy_secret=contrail
-[VIRTUAL-HOST-INTERFACE]
-name=vhost0
-ip=1.1.8.1/24
-physical_interface=eth0
-compute_node_address=1.1.8.1
-gateway=1.1.8.254
-[SERVICE-INSTANCE]
-netns_command=/usr/bin/opencontrail-vrouter-netns
-docker_command=/usr/bin/opencontrail-vrouter-docker
-[HYPERVISOR]
-type = kvm
-[FLOWS]
-fabric_snat_hash_table_size = 4096
-[SESSION]
-slo_destination = collector
-sample_destination = collector`
-
-var vrouterCniConfig = `{
-  "cniVersion": "0.3.1",
-  "contrail" : {
-      "cluster-name"  : "test-cluster",
-      "meta-plugin"   : "multus",
-      "vrouter-ip"    : "127.0.0.1",
-      "vrouter-port"  : 9091,
-      "config-dir"    : "/var/lib/contrail/ports/vm",
-      "poll-timeout"  : 5,
-      "poll-retries"  : 15,
-      "log-file"      : "/var/log/contrail/cni/opencontrail.log",
-      "log-level"     : "4"
-  },
-  "name": "contrail-k8s-cni",
-  "type": "contrail-k8s-cni"
-}`
 
 var devicemanagerWithFabricConfig = `[DEFAULTS]
 host_ip=2.2.2.2
